Fix: decoded raster layers update via setProps
