+++ applications/docs/src/stories/playground/decoded-raster-layer/default.stories.tsx
@@ -1,5 +1,5 @@
 
-import React, { useCallback, useMemo, useState } from 'react';
+import React, { useCallback, useEffect, useMemo, useState } from 'react';
 import { Story } from '@storybook/react/types-6-0';
 // Layer manager
 import { LayerManager, Layer, LayerProps } from '@vizzuality/layer-manager-react';
@@ -107,7 +107,8 @@ const Template: Story<LayerProps> = (args: any) => {
               tile,
               visible,
               opacity,
-              decodeParams,
+              decodeParams: dParams,
+              decodeFunction: dFunction,
             } = sl;
 
             const {
@@ -131,11 +132,12 @@ const Template: Story<LayerProps> = (args: any) => {
                 zoom: z,
                 visible,
                 opacity,
-                decodeParams: decodeParams || {
-                  startYear: 2001,
-                  endYear: 2017,
+                decodeParams: dParams,
+                decodeFunction: dFunction,
+                updateTriggers: {
+                  decodeParams: dParams,
+                  decodeFunction: dFunction,
                 },
-                decodeFunction
               });
             }
             return null;
@@ -151,11 +153,19 @@ const Template: Story<LayerProps> = (args: any) => {
     setViewport(vw);
   }, []);
 
+  useEffect(() => {
+    const [layer] = DECK_LAYERS;
+    if (layer && typeof layer.setProps === 'function') {
+      layer.setProps({
+        decodeParams,
+        decodeFunction,
+      });
+    }
+  }, [decodeParams, decodeFunction])
+
+
   return (
     <div
-      key={JSON.stringify({
-        id, tileUrl, decodeFunction, decodeParams
-      })}
       style={{
         position: 'relative',
         width: '100%',
