+++ applications/docs/src/stories/playground/decoded-raster-layer/glad.stories.tsx
@@ -1,5 +1,5 @@
 
-import React, { useCallback, useMemo, useState } from 'react';
+import React, { useCallback, useEffect, useMemo, useState } from 'react';
 import { Story } from '@storybook/react/types-6-0';
 // Layer manager
 import { LayerManager, Layer, LayerProps } from '@vizzuality/layer-manager-react';
@@ -83,7 +83,7 @@ export default {
 };
 
 const Template: Story<LayerProps> = (args: any) => {
-  const { id, tileUrl, decodeFunction, decodeParams } = args;
+  const { tileUrl, decodeFunction, decodeParams } = args;
 
   const minZoom = 2;
   const maxZoom = 20;
@@ -101,13 +101,17 @@ const Template: Story<LayerProps> = (args: any) => {
           tileSize: 256,
           visible: true,
           refinementStrategy: 'no-overlap',
+          decodeFunction,
+          decodeParams,
           renderSubLayers: (sl) => {
             const {
               id: subLayerId,
               data,
               tile,
               visible,
-              opacity
+              opacity,
+              decodeFunction: dFunction,
+              decodeParams: dParams
             } = sl;
 
             const {
@@ -131,11 +135,12 @@ const Template: Story<LayerProps> = (args: any) => {
                 zoom: z,
                 visible,
                 opacity,
-                decodeParams: decodeParams || {
-                  startYear: 2001,
-                  endYear: 2017,
+                decodeParams: dParams,
+                decodeFunction: dFunction,
+                updateTriggers: {
+                  decodeParams: dParams,
+                  decodeFunction: dFunction,
                 },
-                decodeFunction
               });
             }
             return null;
@@ -151,11 +156,18 @@ const Template: Story<LayerProps> = (args: any) => {
     setViewport(vw);
   }, []);
 
+  useEffect(() => {
+    const [layer] = DECK_LAYERS;
+    if (layer && typeof layer.setProps === 'function') {
+      layer.setProps({
+        decodeParams,
+        decodeFunction
+      });
+    }
+  }, [decodeParams, decodeFunction])
+
   return (
     <div
-      key={JSON.stringify({
-        id, tileUrl, decodeFunction, decodeParams
-      })}
       style={{
         position: 'relative',
         width: '100%',
